[Core] Addresses potential itimer race condition
When `TimerManager::Core::requestTimeout` updates itimer expiration time, it is
possible that the old timer would concurrently expire and set the pending
signal flag. In this case, when we unblock the signal, the signal handler would
be triggered with a wrong `currentTarget` set.

Instead, we proactively stop the itimer before (re)starting it. If its current
expiration is near, we simply wait out the remaining time. Otherwise, we can
safely stop it by setting a zero value timer.